Refactor PHPCourse, and JokeRules
