+++ Project_OLP_Rest.Test/Tests/RegexTests.cs
@@ -12,23 +12,25 @@ namespace Project_OLP_Rest.Test.Tests
     [TestClass]
     public class RegexTests
     {
-        private IEnumerable<BotRule> GreetingBotRules = new GreetingsRules().Rules;
+        private IEnumerable<BotRule> _greetingBotRules;
         private IEnumerable<BotRule> _goodByeBotRules;
         private IEnumerable<BotRule> _errorBotRules;
-        private List<BotRule> JokeBotRules = JokeRules.rules;
+        private IEnumerable<BotRule> _jokeBotRules;
 
         private RestChatBot chatBot;
 
         public RegexTests()
         {
+            _greetingBotRules = new GreetingsRules().Rules;
             _goodByeBotRules = new GoodbyeRuleSet().Rules;
             _errorBotRules = new ErrorRuleSet().Rules;
+            _jokeBotRules = new JokeRulesSet().Rules;
         }
 
         [TestMethod]
         public void BotTest_Greeting()
         {
-            chatBot = new RestChatBot(GreetingBotRules);
+            chatBot = new RestChatBot(_greetingBotRules);
             string Message = "hi";
             
             ChatSessionInterface session = new RestChatSession();
@@ -102,9 +104,9 @@ namespace Project_OLP_Rest.Test.Tests
         [TestMethod]
         public void BotTest_GetJoke()
         {
-            chatBot = new RestChatBot(JokeBotRules);
+            chatBot = new RestChatBot(_jokeBotRules);
             string Message = "tell me a joke";
-            List<string> jokes = JokeRules.jokeList;
+            List<string> jokes = JokeRulesSet.jokeList;
             ChatSessionInterface session = new RestChatSession();
             string answer = chatBot.FindAnswer(session, Message);
             Console.WriteLine(answer);
